Created HeightScreen and WeightScreen

+++ core/src/main/java/com/h_fahmy/core/domain/preferences/Preferences.kt
@@ -16,7 +16,7 @@ interface Preferences {
 
     suspend fun saveHeight(height: Int)
 
-    suspend fun saveWeight(weight: Int)
+    suspend fun saveWeight(weight: Float)
 
     suspend fun saveActivityLevel(activityLevel: ActivityLevel)
